requirements.txt
anyascii==0.3.2
asgiref==3.6.0
bcp47==0.0.4
beautifulsoup4==4.11.2
boto3==1.26.129
botocore==1.29.129
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-tailwind==0.5.0
cryptography==39.0.1
Django==4.1.9
django-appconf==1.0.5
django-browser-reload==1.8.0
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-enumchoicefield==3.0.1
django-extensions==3.2.1
django-filter==22.1
django-ipware==5.0.0
django-js-asset==2.0.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
django-storages==1.13.2
django-taggit==3.1.0
django-tailwind==3.5.0
django-treebeard==4.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-recaptcha==0.2.0
draftjs-exporter==2.1.7
drf-yasg==1.21.5
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
l18n==2021.3
MarkupSafe==2.1.2
openpyxl==3.1.2
packaging==23.1
phonenumbers==8.13.14
Pillow==9.5.0
polib==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tailwind==3.1.5b0
telepath==0.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uWSGI==2.0.20
wagtail==4.1.1
wagtail-cache==2.2.0
wagtail-color-panel==1.4.0
wagtail-favicon==0.2.0
wagtail-generic-chooser==0.5.1
wagtail-localize==1.5
wagtail-metadata==4.0.3
wagtail-robots==0.4.0
wagtailsvg==0.0.37
wagtailtinymce==4.2.1.5
wagtailvideos==4.1.0
webencodings==0.5.1
whitenoise==6.4.0
Willow==1.4.1
